Fill outline triangulated reorganized.
