+++ include/mbgl/shaders/mtl/common.hpp
@@ -65,13 +65,6 @@ struct alignas(16) LineUBO {
     float pad1, pad2, pad3;
 };
 
-struct alignas(16) LineBasicUBO {
-    float4x4 matrix;
-    float2 units_to_pixels;
-    float ratio;
-    float pad;
-};
-
 struct alignas(16) LineGradientUBO {
     float4x4 matrix;
     float ratio;
@@ -86,13 +79,6 @@ struct alignas(16) LinePropertiesUBO {
     float offset;
     float width;
     float pad1, pad2, pad3;
-};
-
-struct alignas(16) LineBasicPropertiesUBO {
-    float4 color;
-    float opacity;
-    float width;
-    float pad1, pad2;
 };
 
 struct alignas(16) LineGradientPropertiesUBO {
