+++ include/mbgl/shaders/shader_defines.hpp
@@ -36,6 +36,7 @@ static constexpr auto maxUBOCountPerShader = std::max({static_cast<size_t>(backg
                                                        static_cast<size_t>(fillOutlineUBOCount),
                                                        static_cast<size_t>(fillPatternUBOCount),
                                                        static_cast<size_t>(fillOutlinePatternUBOCount),
+                                                       static_cast<size_t>(fillOutlineTriangulatedUBOCount),
                                                        static_cast<size_t>(fillExtrusionUBOCount),
                                                        static_cast<size_t>(heatmapUBOCount),
                                                        static_cast<size_t>(heatmapTextureUBOCount),
@@ -45,7 +46,6 @@ static constexpr auto maxUBOCountPerShader = std::max({static_cast<size_t>(backg
                                                        static_cast<size_t>(lineGradientUBOCount),
                                                        static_cast<size_t>(linePatternUBOCount),
                                                        static_cast<size_t>(lineSDFUBOCount),
-                                                       static_cast<size_t>(lineBasicUBOCount),
                                                        static_cast<size_t>(rasterUBOCount),
                                                        static_cast<size_t>(symbolUBOCount)});
