+++ include/mbgl/shaders/shader_source.hpp
@@ -17,19 +17,19 @@ enum class BuiltIn {
     DebugShader,
     FillShader,
     FillOutlineShader,
-    LineGradientShader,
-    LinePatternShader,
-    LineSDFShader,
-    LineShader,
-    LineBasicShader,
     FillPatternShader,
     FillOutlinePatternShader,
+    FillOutlineTriangulatedShader,
     FillExtrusionShader,
     FillExtrusionPatternShader,
     HeatmapShader,
     HeatmapTextureShader,
     HillshadePrepareShader,
     HillshadeShader,
+    LineGradientShader,
+    LinePatternShader,
+    LineSDFShader,
+    LineShader,
     RasterShader,
     SymbolIconShader,
     SymbolSDFIconShader,
