+++ src/mbgl/renderer/layers/fill_layer_tweaker.hpp
@@ -24,12 +24,14 @@ private:
     gfx::UniformBufferPtr fillOutlinePropsUniformBuffer;
     gfx::UniformBufferPtr fillPatternPropsUniformBuffer;
     gfx::UniformBufferPtr fillOutlinePatternPropsUniformBuffer;
+    gfx::UniformBufferPtr fillOutlineTriangulatedUniformBuffer;
 
     // Only run each update function once per property update
     bool fillUniformBufferUpdated = true;
     bool fillOutlineUniformBufferUpdated = true;
     bool fillPatternUniformBufferUpdated = true;
     bool fillOutlinePatternUniformBufferUpdated = true;
+    bool fillOutlineTriangulatedUniformBufferUpdated = true;
 };
 
 } // namespace mbgl
