+++ src/mbgl/renderer/layers/render_fill_layer.cpp
@@ -30,7 +30,6 @@
 #include <mbgl/renderer/update_parameters.hpp>
 #include <mbgl/shaders/fill_layer_ubo.hpp>
 #include <mbgl/shaders/shader_program_base.hpp>
-#include <mbgl/shaders/line_layer_ubo.hpp>
 #endif
 
 namespace mbgl {
@@ -42,12 +41,12 @@ namespace {
 
 #if MLN_DRAWABLE_RENDERER
 constexpr auto FillShaderName = "FillShader";
-#if MLN_TRIANGULATE_FILL_OUTLINES
-constexpr auto FillOutlineTriangulatedShaderName = "LineBasicShader";
-#endif
 constexpr auto FillOutlineShaderName = "FillOutlineShader";
 constexpr auto FillPatternShaderName = "FillPatternShader";
 constexpr auto FillOutlinePatternShaderName = "FillOutlinePatternShader";
+#if MLN_TRIANGULATE_FILL_OUTLINES
+constexpr auto FillOutlineTriangulatedShaderName = "FillOutlineTriangulatedShader";
+#endif
 #endif // MLN_DRAWABLE_RENDERER
 
 inline const FillLayer::Impl& impl_cast(const Immutable<style::Layer::Impl>& impl) {
@@ -104,49 +103,6 @@ bool RenderFillLayer::hasTransition() const {
 bool RenderFillLayer::hasCrossfade() const {
     return getCrossfade<FillLayerProperties>(evaluatedProperties).t != 1;
 }
-
-#if MLN_TRIANGULATE_FILL_OUTLINES
-void RenderFillLayer::markLayerRenderable(bool willRender, UniqueChangeRequestVec& changes) {
-    RenderLayer::markLayerRenderable(willRender, changes);
-    if (outlineLayerGroup) {
-        activateLayerGroup(outlineLayerGroup, willRender, changes);
-    }
-}
-
-void RenderFillLayer::layerRemoved(UniqueChangeRequestVec& changes) {
-    RenderLayer::layerRemoved(changes);
-    if (outlineLayerGroup) {
-        activateLayerGroup(outlineLayerGroup, false, changes);
-    }
-}
-
-void RenderFillLayer::layerIndexChanged(int32_t newLayerIndex, UniqueChangeRequestVec& changes) {
-    RenderLayer::layerIndexChanged(newLayerIndex, changes);
-
-    changeLayerIndex(outlineLayerGroup, newLayerIndex, changes);
-}
-
-std::size_t RenderFillLayer::removeAllDrawables() {
-    auto count = RenderLayer::removeAllDrawables();
-    if (outlineLayerGroup) {
-        auto count2 = outlineLayerGroup->clearDrawables();
-        stats.drawablesRemoved += count2;
-        count += count2;
-    }
-    return count;
-}
-
-std::size_t RenderFillLayer::removeTile(RenderPass renderPass, const OverscaledTileID& tileID) {
-    RenderLayer::removeTile(renderPass, tileID);
-    if (const auto tileGroup = static_cast<TileLayerGroup*>(outlineLayerGroup.get())) {
-        const auto n = tileGroup->removeDrawables(renderPass, tileID).size();
-        stats.drawablesRemoved += n;
-        return n;
-    }
-    return 0;
-}
-
-#endif // MLN_TRIANGULATE_FILL_OUTLINES
 
 #if MLN_LEGACY_RENDERER
 void RenderFillLayer::render(PaintParameters& parameters) {
@@ -338,60 +294,6 @@ bool RenderFillLayer::queryIntersectsFeature(const GeometryCoordinates& queryGeo
 
 #if MLN_DRAWABLE_RENDERER
 
-#if MLN_TRIANGULATE_FILL_OUTLINES
-class OutlineDrawableTweaker : public gfx::DrawableTweaker {
-public:
-    OutlineDrawableTweaker(Color color_, float opacity_)
-        : color(color_),
-          opacity(opacity_) {}
-    ~OutlineDrawableTweaker() override = default;
-
-    void init(gfx::Drawable&) override {};
-
-    virtual void execute(gfx::Drawable& drawable, const PaintParameters& parameters) override {
-        if (!drawable.getTileID().has_value()) {
-            return;
-        }
-
-        const UnwrappedTileID tileID = drawable.getTileID()->toUnwrapped();
-        const auto zoom = parameters.state.getZoom();
-        auto& uniforms = drawable.mutableUniformBuffers();
-
-        {
-            const auto matrix = LayerTweaker::getTileMatrix(
-                tileID, parameters, {{0, 0}}, style::TranslateAnchorType::Viewport, false, false, drawable, false);
-
-            const shaders::LineBasicUBO lineBasicUBO{
-                /*matrix = */ util::cast<float>(matrix),
-                /*units_to_pixels = */ {1.0f / parameters.pixelsToGLUnits[0], 1.0f / parameters.pixelsToGLUnits[1]},
-                /*ratio = */ 1.0f / tileID.pixelsToTileUnits(1.0f, zoom),
-                0};
-            parameters.context.emplaceOrUpdateUniformBuffer(lineUniformBuffer, &lineBasicUBO);
-        }
-        uniforms.set(idLineBasicUBO, lineUniformBuffer);
-
-        if (!linePropertiesUniformBuffer) {
-            const shaders::LineBasicPropertiesUBO linePropertiesUBO{/*color =*/color,
-                                                                    /*opacity =*/opacity,
-                                                                    /*width =*/1.f,
-                                                                    0,
-                                                                    0};
-            parameters.context.emplaceOrUpdateUniformBuffer(linePropertiesUniformBuffer, &linePropertiesUBO);
-        }
-        if (!uniforms.get(idLineBasicPropertiesUBO)) {
-            uniforms.set(idLineBasicPropertiesUBO, linePropertiesUniformBuffer);
-        }
-    };
-
-private:
-    Color color;
-    float opacity;
-
-    gfx::UniformBufferPtr lineUniformBuffer;
-    gfx::UniformBufferPtr linePropertiesUniformBuffer;
-};
-#endif // MLN_TRIANGULATE_FILL_OUTLINES
-
 void RenderFillLayer::update(gfx::ShaderRegistry& shaders,
                              gfx::Context& context,
                              const TransformState& state,
@@ -414,12 +316,6 @@ void RenderFillLayer::update(gfx::ShaderRegistry& shaders,
     auto* fillTileLayerGroup = static_cast<TileLayerGroup*>(layerGroup.get());
 
 #if MLN_TRIANGULATE_FILL_OUTLINES
-    // Set up a layer group for outlines
-    if (!outlineLayerGroup) {
-        outlineLayerGroup = context.createTileLayerGroup(layerIndex, /*initialCapacity=*/64, getID());
-        activateLayerGroup(outlineLayerGroup, isRenderable, changes);
-    }
-    auto* outlineTileLayerGroup = static_cast<TileLayerGroup*>(outlineLayerGroup.get());
     if (!outlineTriangulatedShaderGroup) {
         outlineTriangulatedShaderGroup = shaders.getShaderGroup(std::string(FillOutlineTriangulatedShaderName));
     }
@@ -460,21 +356,11 @@ void RenderFillLayer::update(gfx::ShaderRegistry& shaders,
         builder.setEnableStencil(true);
     };
 
-#if MLN_TRIANGULATE_FILL_OUTLINES
-    for (auto layer : {fillTileLayerGroup, outlineTileLayerGroup}) {
-        stats.drawablesRemoved += layer->removeDrawablesIf([&](gfx::Drawable& drawable) {
-            // If the render pass has changed or the tile has dropped out of the cover set, remove it.
-            const auto& tileID = drawable.getTileID();
-            return tileID && !hasRenderTile(*tileID);
-        });
-    }
-#else
     stats.drawablesRemoved += fillTileLayerGroup->removeDrawablesIf([&](gfx::Drawable& drawable) {
         // If the render pass has changed or the tile has dropped out of the cover set, remove it.
         const auto& tileID = drawable.getTileID();
         return tileID && !hasRenderTile(*tileID);
     });
-#endif
 
     fillTileLayerGroup->setStencilTiles(renderTiles);
 
@@ -690,7 +576,7 @@ void RenderFillLayer::update(gfx::ShaderRegistry& shaders,
             }()
                 : nullptr;
 
-            auto createOutline = [&](auto& builder, Color color, float opacity) {
+            auto createOutlineTriangulated = [&](auto& builder) {
                 if (doOutline && builder && bucket.sharedLineIndexes->elements()) {
                     builder->setShader(outlineTriangulatedShader);
                     builder->setRawVertices({}, bucket.lineVertices.elements(), gfx::AttributeDataType::Short2);
@@ -717,14 +603,12 @@ void RenderFillLayer::update(gfx::ShaderRegistry& shaders,
                                          bucket.lineSegments.data(),
                                          bucket.lineSegments.size());
 
-                    auto tweaker = std::make_shared<OutlineDrawableTweaker>(color, opacity);
-
                     // finish
                     builder->flush(context);
                     for (auto& drawable : builder->clearDrawables()) {
                         drawable->setTileID(tileID);
-                        drawable->addTweaker(tweaker);
-                        outlineTileLayerGroup->addDrawable(renderPass, tileID, std::move(drawable));
+                        drawable->setType(static_cast<size_t>(FillVariant::FillOutlineTriangulated));
+                        fillTileLayerGroup->addDrawable(renderPass, tileID, std::move(drawable));
                         ++stats.drawablesAdded;
                     }
                 }
@@ -808,9 +692,7 @@ void RenderFillLayer::update(gfx::ShaderRegistry& shaders,
             if (doOutline && outlineBuilder) {
                 if (!dataDrivenOutline) {
                     outlineBuilder->setSubLayerIndex(unevaluated.get<FillOutlineColor>().isUndefined() ? 2 : 0);
-                    createOutline(outlineBuilder,
-                                  evaluated.get<FillOutlineColor>().constantOr(FillOutlineColor::defaultValue()),
-                                  evaluated.get<FillOpacity>().constantOr(FillOpacity::defaultValue()));
+                    createOutlineTriangulated(outlineBuilder);
                 } else {
                     if (bucket.sharedBasicLineIndexes->elements()) {
                         outlineBuilder->setShader(outlineShader);
